Persistent flags are added to the environment cli
